requirements.txt
numpy
pandas
nibabel 
p-tqdm
pytorch-lightning
timm
torchmetrics
scipy
sklearn